Added script deletion support in the editor
However, there is no frontend.

On the main process, it now listens for "delete" and when it gets it
with a script name, it attempts to delete it.

If that succeeds, the main process will tell the frontend that it is
deleted, and will cause the editor to reload the script selection
drop-down and to load the first file in it.

It is currently assumed that the file deleted is the current one, though
a future enhancement should stop the entire editor from reloading if the
file you are currently editing still exists.