Added universal select list function in hells.js
Now, if you want to provide a user with a lost of choices, you can do
this by appendChilding the result of a makeList call.

For example, to add a list to an element in the variable "thing", you
can simply:

  thing.appendChild(makeList(["Oh", "boy"], val => console.log(val)));

This will give an ordered list containing "Oh" and "boy" that, when
clicked by the user, will be output with console.log() and the entire
list will disappear.

There is currently no styling though.